+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#141](https://github.com/ruby-grape/grape-swagger-rails/pull/141): Replace swagger ui 2.x with swaggeruibundle v5 standalone integration - [@moskvin](https://github.com/moskvin).
 * [#142](https://github.com/ruby-grape/grape-swagger-rails/pull/142): Convert view to haml and extract javascript to typescript - [@moskvin](https://github.com/moskvin).
 * [#143](https://github.com/ruby-grape/grape-swagger-rails/pull/143): Add swagger_ui_config pass-through to swaggeruibundle - [@moskvin](https://github.com/moskvin).
+* [#144](https://github.com/ruby-grape/grape-swagger-rails/pull/144): Add support for multiple api specifications in swagger ui - [@moskvin](https://github.com/moskvin).
 * Your contribution here.
 
 ### 0.7.0 (2025/09/16)